# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/audit-logs": {
            "get": {
                "description": "根据操作者ID分页查询审计日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "审计"
                ],
                "summary": "按操作者查询审计日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作者ID",
                        "name": "operator_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "缺少操作者ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
        "/audit-logs/target": {
            "get": {
                "description": "根据目标类型和ID分页查询审计日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "审计"
                ],
                "summary": "按目标查询审计日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标类型",
                        "name": "target_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目标ID",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "缺少参数",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
        "/audit-logs/time": {
            "get": {
                "description": "根据时间范围分页查询审计日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "审计"
                ],
                "summary": "按时间范围查询审计日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lyj404_gin-api-template_domain_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.PaginationResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "type": "string"
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}